Added Canvas example
The canvas example is in examples under the file name
"canvas_image_gen". If it's too long we can change it. I'm not too
attached. Play around with it a bit. Right now it's set to loop and
resize so if you'd like to not have it do either just comment the lines
out.